+++ cAutoPACK/src/IngredientsDispatcher.h
@@ -15,15 +15,15 @@ public: //Temporary public, should be private
     std::vector<Ingredient*> activeIngr0;
     std::vector<Ingredient*> activeIngr12;
 
-    std::vector<float> normalizedPriorities;
-    float vRangeStart;
+    std::vector<double> normalizedPriorities;
+    double vRangeStart;
 
-    std::vector<float> thresholdPriorities; 
+    std::vector<double> thresholdPriorities; 
 
     std::vector<Ingredient> ingredients;//the list of sphere ingredient to pack
 
 
-    std::uniform_real_distribution<float> uniform;
+    std::uniform_real_distribution<double> uniform;
     std::default_random_engine generator;
 
 public:
